Changed from live update to local database.

+++ app/src/main/java/com/boamfa/workout/classes/TrackDayExerciseSet.java
@@ -6,11 +6,11 @@ import java.io.Serializable;
  * Created by bogdan on 10/12/15.
  */
 public class TrackDayExerciseSet implements Serializable {
-    public int id;
+    public long id;
     public int reps;
     public double weight;
 
-    public TrackDayExerciseSet(int id, int reps, double weight) {
+    public TrackDayExerciseSet(long id, int reps, double weight) {
         this.id = id;
         this.reps = reps;
         this.weight = weight;
